# backend/tests/openapi_snapshot.json
{
  "components": {
    "schemas": {
...
      "EventSourceConfigOut": {
        "additionalProperties": false,
        "properties": {
          "endpoints": {
            "items": {
              "$ref": "#/components/schemas/EventSourceEndpointOut"
            },
            "title": "Endpoints",
            "type": "array"
          },
          "grafana": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EventSourceGrafanaOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "last_test_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Test Status"
          },
          "max_parallel_endpoints": {
            "title": "Max Parallel Endpoints",
            "type": "integer"
          },
          "poll_interval_seconds": {
            "title": "Poll Interval Seconds",
            "type": "integer"
          },
          "resolution_grace_seconds": {
            "title": "Resolution Grace Seconds",
            "type": "integer"
          },
          "safe_error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Safe Error Code"
          },
          "tested_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tested At"
          },
          "thanos": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EventSourceThanosOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "watchdog_alertname": {
            "title": "Watchdog Alertname",
            "type": "string"
          },
          "watchdog_enabled": {
            "title": "Watchdog Enabled",
            "type": "boolean"
          },
          "watchdog_identity_label": {
            "title": "Watchdog Identity Label",
            "type": "string"
          },
          "watchdog_missing_after_seconds": {
            "title": "Watchdog Missing After Seconds",
            "type": "integer"
          }
        },
        "required": [
          "poll_interval_seconds",
          "resolution_grace_seconds",
          "max_parallel_endpoints",
          "watchdog_enabled",
          "watchdog_alertname",
          "watchdog_identity_label",
          "watchdog_missing_after_seconds",
          "last_test_status",
          "tested_at",
          "safe_error_code",
          "endpoints",
          "thanos"
        ],
        "title": "EventSourceConfigOut",
        "type": "object"
      },
...
      "EventSourceEndpointOut": {
        "additionalProperties": false,
        "properties": {
          "auth_type": {
            "title": "Auth Type",
            "type": "string"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "last_test": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EventSourceEndpointLastTestOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "position": {
            "title": "Position",
            "type": "integer"
          },
          "secret_configured": {
            "title": "Secret Configured",
            "type": "boolean"
          },
          "url": {
            "title": "Url",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "position",
          "url",
          "enabled",
          "auth_type",
          "username",
          "secret_configured",
          "last_test"
        ],
        "title": "EventSourceEndpointOut",
        "type": "object"
      },
...
      "EventSourceGrafanaIn": {
        "additionalProperties": false,
        "description": "The optional Grafana address. An empty URL removes the configuration.\n\nNo `auth_type` and no `username`: Grafana here is Bearer-or-anonymous, and\nan internal instance configured for anonymous read is a normal deployment,\nnot a misconfigured one. An absent credential *is* the anonymous case.\n\n**Plain `http` and a private address are both accepted.** Grafana is a\nread-only monitoring source living on the internal network, not an outbound\ntarget — applying the outbound address rules here would reject the ordinary\nconfiguration (CAP-13.1a).",
        "properties": {
          "secret": {
            "$ref": "#/components/schemas/SecretUpdate"
          },
          "timeout_seconds": {
            "default": 15,
            "maximum": 120.0,
            "minimum": 1.0,
            "title": "Timeout Seconds",
            "type": "integer"
          },
          "url": {
            "default": "",
            "maxLength": 2048,
            "title": "Url",
            "type": "string"
          }
        },
        "title": "EventSourceGrafanaIn",
        "type": "object"
      },
...
      "EventSourceOut": {
        "additionalProperties": false,
        "properties": {
          "archived_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived At"
          },
          "config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EventSourceConfigOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "lifecycle_state": {
            "title": "Lifecycle State",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "type",
          "name",
          "lifecycle_state",
          "status",
          "version",
          "config",
          "created_at",
          "updated_at",
          "archived_at"
        ],
        "title": "EventSourceOut",
        "type": "object"
      },
...
      "GrafanaDashboardSummaryOut": {
        "additionalProperties": false,
        "properties": {
          "folder_title": {
            "default": "",
            "title": "Folder Title",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "uid": {
            "title": "Uid",
            "type": "string"
          }
        },
        "required": [
          "uid",
          "title"
        ],
        "title": "GrafanaDashboardSummaryOut",
        "type": "object"
      },
      "GrafanaImportConfirmIn": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/GrafanaImportItemIn"
            },
            "maxItems": 200,
            "title": "Items",
            "type": "array"
          }
        },
        "title": "GrafanaImportConfirmIn",
        "type": "object"
      },
      "GrafanaImportConfirmOut": {
        "additionalProperties": false,
        "properties": {
          "created_template_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Created Template Ids",
            "type": "array"
          },
          "updated_template_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Updated Template Ids",
            "type": "array"
          }
        },
        "required": [
          "created_template_ids",
          "updated_template_ids"
        ],
        "title": "GrafanaImportConfirmOut",
        "type": "object"
      },
      "GrafanaImportItemIn": {
        "additionalProperties": false,
        "description": "A ticked candidate as the browser finally submitted it.\n\n`final_promql` is the literal the user read on screen. The server does not\nre-fetch the dashboard to rebuild it: that would look more careful and be\nless so, because a definition changed between preview and confirmation would\nleave the stored query different from the approved one (review Q4).",
        "properties": {
          "display_unit": {
            "default": "",
            "maxLength": 40,
            "title": "Display Unit",
            "type": "string"
          },
          "enabled": {
            "default": false,
            "title": "Enabled",
            "type": "boolean"
          },
          "final_promql": {
            "maxLength": 4096,
            "minLength": 1,
            "title": "Final Promql",
            "type": "string"
          },
          "imported_promql": {
            "default": "",
            "maxLength": 4096,
            "title": "Imported Promql",
            "type": "string"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "order": {
            "default": 0,
            "maximum": 10000.0,
            "minimum": 0.0,
            "title": "Order",
            "type": "integer"
          },
          "origin": {
            "$ref": "#/components/schemas/ImportOriginIn"
          },
          "required_labels": {
            "default": [],
            "items": {
              "type": "string"
            },
            "maxItems": 16,
            "title": "Required Labels",
            "type": "array"
          },
          "template_id": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Id"
          }
        },
        "required": [
          "final_promql",
          "name",
          "origin"
        ],
        "title": "GrafanaImportItemIn",
        "type": "object"
      },
      "GrafanaImportPreviewIn": {
        "additionalProperties": false,
        "properties": {
          "dashboard_uid": {
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Dashboard Uid",
            "type": "string"
          }
        },
        "required": [
          "dashboard_uid"
        ],
        "title": "GrafanaImportPreviewIn",
        "type": "object"
      },
      "GrafanaImportPreviewOut": {
        "additionalProperties": false,
        "properties": {
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/ImportCandidateOut"
            },
            "title": "Candidates",
            "type": "array"
          },
          "dashboard_title": {
            "title": "Dashboard Title",
            "type": "string"
          },
          "dashboard_uid": {
            "title": "Dashboard Uid",
            "type": "string"
          },
          "enabled_template_count": {
            "title": "Enabled Template Count",
            "type": "integer"
          },
          "max_auxiliary_curves": {
            "title": "Max Auxiliary Curves",
            "type": "integer"
          },
          "probed": {
            "title": "Probed",
            "type": "boolean"
          }
        },
        "required": [
          "dashboard_uid",
          "dashboard_title",
          "candidates",
          "probed",
          "enabled_template_count",
          "max_auxiliary_curves"
        ],
        "title": "GrafanaImportPreviewOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SecretUpdate": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "enum": [
              "KEEP",
              "REPLACE",
              "CLEAR"
            ],
            "title": "Action",
            "type": "string"
          },
          "value": {
            "anyOf": [
              {
                "maxLength": 4096,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "action"
        ],
        "title": "SecretUpdate",
        "type": "object"
      },
...
      "TestResultOut": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "transient": {
            "default": false,
            "title": "Transient",
            "type": "boolean"
          }
        },
        "required": [
          "ok",
          "code"
        ],
        "title": "TestResultOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Alert Workbench",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/event-sources/{source_id}/grafana": {
      "put": {
        "operationId": "save_grafana_config_api_event_sources__source_id__grafana_put",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventSourceGrafanaIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventSourceOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Grafana Config"
      }
    },
    "/api/event-sources/{source_id}/grafana/dashboards": {
      "get": {
        "operationId": "list_grafana_dashboards_api_event_sources__source_id__grafana_dashboards_get",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "query",
            "required": false,
            "schema": {
              "default": "",
              "title": "Query",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GrafanaDashboardSummaryOut"
                  },
                  "title": "Response List Grafana Dashboards Api Event Sources  Source Id  Grafana Dashboards Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Grafana Dashboards"
      }
    },
    "/api/event-sources/{source_id}/grafana/import/confirm": {
      "post": {
        "description": "The only writing route. One transaction for the whole batch.",
        "operationId": "confirm_grafana_import_api_event_sources__source_id__grafana_import_confirm_post",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrafanaImportConfirmIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrafanaImportConfirmOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Confirm Grafana Import"
      }
    },
    "/api/event-sources/{source_id}/grafana/import/preview": {
      "post": {
        "description": "Read, parse, diff and check — and write nothing.\n\nThere is no server-side draft to store: the candidates go to the browser and\ncome back as literals on confirmation. That is also why this function\ncontains no `session.add`, `flush` or `commit`, which a test asserts by\ncounting rows before and after.",
        "operationId": "preview_grafana_import_api_event_sources__source_id__grafana_import_preview_post",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrafanaImportPreviewIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrafanaImportPreviewOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview Grafana Import"
      }
    },
    "/api/event-sources/{source_id}/grafana/test": {
      "post": {
        "description": "One read-only listing request. Explicit, and the only thing that opens\nthe import gate.",
        "operationId": "test_grafana_config_api_event_sources__source_id__grafana_test_post",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestResultOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Test Grafana Config"
      }
    },
...
  }
}